+++ src/applications/ivc-champva/10-7959f-1/tests/unit/containers/ConfirmationPage.unit.spec.js
@@ -8,7 +8,7 @@ import { expect } from 'chai';
 import ConfirmationPage from '../../../containers/ConfirmationPage';
 import formConfig from '../../../config/form';
 
-import mockData from '../../fixtures/data/test-data.json';
+import mockData from '../../e2e/fixtures/data/test-data.json';
 
 const subDate = new Date('11/13/2023').toString();
